requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
distro==1.9.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
openai>=1.3.0
pika>=1.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic_core==2.27.2
python-dotenv>=1.0.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
websockets>=11.0.0
pyodbc==5.2.0
aiohttp>=3.8.0
redis>=5.0.0
requests>=2.31.0
